# react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/letters/letter": {
      "post": {
        "tags": [
          "letters"
        ],
        "summary": "Add Next Letter",
        "description": "Create a new letter for a group.\n\nCreates a new letter with default questions if there are no letters\ncurrently in progress or upcoming for the group.\n\nArgs:\n    letter (LetterCreate): Letter creation parameters\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nReturns:\n    Letter: Newly created letter\n\nRaises:\n    ValueError: If there is already a letter in progress or upcoming",
        "operationId": "add_next_letter_letters_letter_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LetterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLetter"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/letters/letters/": {
      "get": {
        "tags": [
          "letters"
        ],
        "summary": "List Letters",
        "description": "List letters for a specific group.\n\nRetrieves a paginated list of letters belonging to the specified group.\n\nArgs:\n    group_api_id (str): API identifier of the group\n    skip (int, optional): Number of records to skip for pagination. Defaults to 0.\n    limit (int, optional): Maximum number of records to return. Defaults to 100.\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nReturns:\n    Sequence[Letter]: List of letters",
        "operationId": "list_letters_letters_letters__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_api_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Api Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicLetter"
                  },
                  "title": "Response List Letters Letters Letters  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/letters/letter/{letter_api_id}": {
      "get": {
        "tags": [
          "letters"
        ],
        "summary": "Read Letter",
        "description": "Retrieve a specific letter by its API identifier.\n\nArgs:\n    letter_api_id (str): API identifier of the letter\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nReturns:\n    Letter: The requested letter\n\nRaises:\n    IDNotFoundException: If letter with given API ID is not found",
        "operationId": "read_letter_letters_letter__letter_api_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "letter_api_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Letter Api Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLetter"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/letters/letters:dashboard": {
      "get": {
        "tags": [
          "letters"
        ],
        "summary": "List Dashboard Letters",
        "description": "List letters for the dashboard view.\n\nRetrieves letters that are upcoming, in progress, or recently completed\n(within the last 8 days) for the current user.\n\nArgs:\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nReturns:\n    dict[str, list[Letter]]: Dictionary containing categorized letters",
        "operationId": "list_dashboard_letters_letters_letters_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLetters"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/letters/letter/{letter_api_id}:edit_letter": {
      "post": {
        "tags": [
          "letters"
        ],
        "summary": "Edit Letter",
        "description": "Update a letter's details.\n\nUpdates the send time of a letter. For upcoming letters, ensures the new\nsend time is after any in-progress letter's send time.\n\nArgs:\n    letter_api_id (str): API identifier of the letter to edit\n    letter (LetterUpdate): Updated letter details\n    req_dep (AuthenticatedRequestDependencies): Request dependencies including database session and auth\n\nReturns:\n    Letter: Updated letter\n\nRaises:\n    AssertionError: If new send time violates timing constraints\n    IDNotFoundException: If letter with given API ID is not found",
        "operationId": "edit_letter_letters_letter__letter_api_id__edit_letter_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "letter_api_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Letter Api Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LetterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLetter"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/letters/letter/{letter_api_id}:add_question": {
      "post": {
        "tags": [
          "letters"
        ],
        "summary": "Add Question",
        "description": "Add a new question to a letter.\n\nCreates and adds a new question to the specified letter. The question can\noptionally be associated with an author.\n\n:param letter_api_id: API identifier of the letter\n:type letter_api_id: str\n:param question: Question creation parameters\n:type question: QuestionCreate\n:param req_dep: Request dependencies including database session and auth\n:type req_dep: AuthenticatedRequestDependencies\n:raises IDNotFoundException: If letter or author with given API ID is not found\n:return: Updated letter with the new question\n:rtype: Letter",
        "operationId": "add_question_letters_letter__letter_api_id__add_question_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "letter_api_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Letter Api Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLetter"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/letters/letter/{letter_api_id}:generate_question": {
      "post": {
        "tags": [
          "letters"
        ],
        "summary": "Generate Question",
        "description": "Generate a question using LLM without saving it.",
        "operationId": "generate_question_letters_letter__letter_api_id__generate_question_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "letter_api_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Letter Api Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateQuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardLetters": {
        "properties": {
          "upcoming": {
            "items": {
              "$ref": "#/components/schemas/PublicLetter"
            },
            "type": "array",
            "title": "Upcoming"
          },
          "in_progress": {
            "items": {
              "$ref": "#/components/schemas/PublicLetter"
            },
            "type": "array",
            "title": "In Progress"
          },
          "recently_completed": {
            "items": {
              "$ref": "#/components/schemas/PublicLetter"
            },
            "type": "array",
            "title": "Recently Completed"
          }
        },
        "type": "object",
        "required": [
          "upcoming",
          "in_progress",
          "recently_completed"
        ],
        "title": "DashboardLetters",
        "description": "Model for the letters dashboard view.\n\nGroups letters by their status for dashboard display.\n\nAttributes:\n    upcoming (list[PublicLetter]): Letters scheduled for the future\n    in_progress (list[PublicLetter]): Currently active letters\n    recently_completed (list[PublicLetter]): Recently finished letters"
      },
      "GenerateQuestionRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "GenerateQuestionRequest"
      },
      "GenerateQuestionResponse": {
        "properties": {
          "generated_text": {
            "type": "string",
            "title": "Generated Text"
          }
        },
        "type": "object",
        "required": [
          "generated_text"
        ],
        "title": "GenerateQuestionResponse"
      },
...
      "GroupUnlinked": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "cycle_length": {
            "type": "integer",
            "title": "Cycle Length"
          }
        },
        "type": "object",
        "required": [
          "name",
          "api_identifier",
          "created_at",
          "cycle_length"
        ],
        "title": "GroupUnlinked",
        "description": "Schema for groups without linked relationships.\n\nInherits all fields from Group but excludes relationship data."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Image": {
        "properties": {
          "s3_url": {
            "type": "string",
            "title": "S3 Url"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          }
        },
        "type": "object",
        "required": [
          "s3_url",
          "media_type"
        ],
        "title": "Image",
        "description": "Schema for image data.\n\nRepresents an image stored in S3 with its media type.\n\nAttributes:\n    s3_url (str): S3 key/path for the image\n    media_type (MediaType): Type of media (image/video)"
      },
...
      "LetterCreate": {
        "properties": {
          "group_api_identifier": {
            "type": "string",
            "title": "Group Api Identifier"
          },
          "send_at": {
            "type": "string",
            "format": "date-time",
            "title": "Send At"
          }
        },
        "type": "object",
        "required": [
          "group_api_identifier",
          "send_at"
        ],
        "title": "LetterCreate",
        "description": "Schema for creating a new letter.\n\nAttributes:\n    group_api_identifier (str): API identifier of the group to create the letter for\n    send_at (AwareDatetime): Scheduled time to send the letter"
      },
      "LetterStatus": {
        "type": "string",
        "enum": [
          "UPCOMING",
          "IN_PROGRESS",
          "SENT"
        ],
        "title": "LetterStatus",
        "description": "Enumeration of possible letter statuses.\nRepresents the different states a letter can be in within the system:\n- UPCOMING: Letter is scheduled but not yet active\n- IN_PROGRESS: Letter is currently active and accepting responses\n- SENT: Letter has been completed and sent"
      },
...
      "LetterUpdate": {
        "properties": {
          "send_at": {
            "type": "string",
            "format": "date-time",
            "title": "Send At"
          }
        },
        "type": "object",
        "required": [
          "send_at"
        ],
        "title": "LetterUpdate",
        "description": "Schema for updating an existing letter.\n\nAttributes:\n    send_at (AwareDatetime): New scheduled time to send the letter"
      },
      "MediaType": {
        "type": "string",
        "enum": [
          "image",
          "video"
        ],
        "title": "MediaType",
        "description": "Enumeration of supported media types.\n\nAttributes:\n    IMAGE: Image files (e.g., jpg, png)\n    VIDEO: Video files"
      },
...
      "PublicLetter": {
        "properties": {
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "status": {
            "$ref": "#/components/schemas/LetterStatus"
          },
          "send_at": {
            "type": "string",
            "format": "date-time",
            "title": "Send At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "group": {
            "$ref": "#/components/schemas/GroupUnlinked"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/PublicQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "responders": {
            "items": {
              "$ref": "#/components/schemas/UserUnlinked"
            },
            "type": "array",
            "title": "Responders"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/UserUnlinked"
            },
            "type": "array",
            "title": "Participants"
          }
        },
        "type": "object",
        "required": [
          "api_identifier",
          "number",
          "status",
          "send_at",
          "created_at",
          "group",
          "questions",
          "responders",
          "participants"
        ],
        "title": "PublicLetter",
        "description": "Public letter model with linked relationships.\n\nA version of the Letter model that includes public information about participants\nand responses.\n\nAttributes:\n    group (GroupUnlinked): Group the letter belongs to\n    questions (list[PublicQuestion]): Questions with public responses\n    responders (list[UserUnlinked]): Users who have responded\n    participants (list[UserUnlinked]): All participants in the letter"
      },
      "PublicQuestion": {
        "properties": {
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "responses": {
            "items": {
              "$ref": "#/components/schemas/ResponseWithParticipant"
            },
            "type": "array",
            "title": "Responses"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserUnlinked"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "question_text",
          "api_identifier",
          "created_at",
          "responses",
          "author"
        ],
        "title": "PublicQuestion",
        "description": "Public question model with linked relationships.\n\nA version of the Question model that includes public response information.\n\nAttributes:\n    responses (list[ResponseWithParticipant]): Public responses with participant info\n    author (Optional[UserUnlinked]): Question author, if available"
      },
      "QuestionCreate": {
        "properties": {
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "author_api_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Api Id"
          }
        },
        "type": "object",
        "required": [
          "question_text",
          "author_api_id"
        ],
        "title": "QuestionCreate",
        "description": "Schema for creating a new question.\n\nAttributes:\n    question_text (str): The text content of the question\n    author_api_id (str | None): API identifier of the question's author, optional"
      },
...
      "ResponseWithParticipant": {
        "properties": {
          "response_text": {
            "type": "string",
            "title": "Response Text"
          },
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "participant": {
            "$ref": "#/components/schemas/UserUnlinked"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/Image"
            },
            "type": "array",
            "title": "Images",
            "description": "Get the list of images associated with this model.\n\nReturns:\n    list[Image]: List of associated images",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "response_text",
          "api_identifier",
          "created_at",
          "participant",
          "images"
        ],
        "title": "ResponseWithParticipant",
        "description": "Response model with participant information and image support.\n\nA version of the Response model that includes participant information\nand support for attached images.\n\nAttributes:\n    participant (UserUnlinked): User who provided the response"
      },
...
      "UserUnlinked": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "api_identifier": {
            "type": "string",
            "title": "Api Identifier"
          },
          "admin": {
            "type": "boolean",
            "title": "Admin"
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "api_identifier",
          "admin"
        ],
        "title": "UserUnlinked",
        "description": "Schema for users without linked relationships.\n\nInherits all fields from User but excludes relationship data."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "login/access-token"
          }
        }
      }
    }
  }
}